requirements.txt
streamlit==1.28.0
langchain==0.0.350
openai==1.3.5
python-dotenv==1.0.0
streamlit-chat==0.1.1
